Move convert_color_to_hex function to utility
