+++ Framework/PythonInterface/mantid/plots/__init__.py
@@ -18,11 +18,11 @@ from matplotlib.projections import register_projection
 from matplotlib.scale import register_scale
 
 from mantid.plots import datafunctions, axesfunctions, axesfunctions3D  # noqa: F401
-from mantid.plots.legend import convert_color_to_hex, LegendProperties  # noqa: F401
+from mantid.plots.legend import LegendProperties  # noqa: F401
 from mantid.plots.datafunctions import get_normalize_by_bin_width  # noqa: F401
 from mantid.plots.scales import PowerScale, SquareScale  # noqa: F401
 from mantid.plots.mantidaxes import MantidAxes, MantidAxes3D, WATERFALL_XOFFSET_DEFAULT, WATERFALL_YOFFSET_DEFAULT  # noqa: F401
-from mantid.plots.utility import (artists_hidden, autoscale_on_update, legend_set_draggable, MantidAxType)  # noqa: F401
+from mantid.plots.utility import artists_hidden, autoscale_on_update, convert_color_to_hex, legend_set_draggable, MantidAxType  # noqa: F401
 
 register_projection(MantidAxes)
 register_projection(MantidAxes3D)
