+++ qt/applications/workbench/workbench/plotting/test/test_figureinteraction.py
@@ -602,7 +602,7 @@ class FigureInteractionTest(unittest.TestCase):
         self.assertEqual(1, self.interactor.redraw_annotations.call_count)
 
     def test_toggle_normalisation_on_contour_plot_maintains_contour_line_colour(self):
-        from mantid.plots.legend import convert_color_to_hex
+        from mantid.plots.utility import convert_color_to_hex
         ws = CreateWorkspace(DataX=[1, 2, 3, 4, 2, 4, 6, 8], DataY=[2] * 8, NSpec=2, OutputWorkspace="test_ws")
         fig = plot_contour([ws])
