+++ qt/python/mantidqt/mantidqt/widgets/plotconfigdialog/colorselector.py
@@ -6,7 +6,7 @@
 # SPDX - License - Identifier: GPL - 3.0 +
 #  This file is part of the mantid workbench.
 
-from mantid.plots.legend import convert_color_to_hex
+from mantid.plots.utility import convert_color_to_hex
 from matplotlib import rcParams
 from qtpy.QtCore import QRegExp
 from qtpy.QtGui import QColor, QRegExpValidator
